apps/feed-service-python/requirements.txt
grpcio
grpcio-tools
protobuf
pylint
black
